extract shared cairo-starknet-syscalls crate
Both cairo-native and sierra-emu now re-export the StarknetSyscallHandler
trait and supporting types (U256, Block/Tx/Execution Info v{1,2,3},
ResourceBounds, Secp256{k1,r1}Point) from the new
`cairo-starknet-syscalls` crate. The crate is workspace-local at the
top level (sibling to `test_utils/`).

The two crates' types are now nominally identical, so a single handler
impl can drive both runtimes without an adapter / bridge.

Notes:
- `cairo-starknet-syscalls` rather than `cairo-native-syscalls`: the
  crate's content doesn't depend on cairo-native semantically; the
  cairo-native prefix would imply ownership it doesn't have.
- StubSyscallHandler's `unimplemented!()` stubs now carry a one-line
  rationale each (deploy needs constructor execution, call_contract
  needs address-to-class resolution, send_message_to_l1 needs an L1
  queue, etc.), plus a scope doc-comment on the type itself.
- sierra-emu's previous inherent into_value/from_value methods (which
  depend on sierra-emu's `Value` enum and can't follow the moved types)
  become free functions in the new sub-module
  `sierra_emu::starknet::value_conv`, each with a doc comment.
- `cheatcode` trait method's default impl returns an error felt instead
  of panicking with unimplemented!(). On the sierra-emu path, a contract
  invoking the cheatcode libfunc against a handler that didn't override
  it would otherwise unwind through the VM and abort the host;
  production handlers (e.g. blockifier's NativeSyscallHandler) don't
  implement cheatcode, so the default is what runs.
- The per-type files under debug_utils/sierra-emu/src/starknet/* are
  deleted now that the struct definitions live in the shared crate.
