+++ Cargo.toml
@@ -3,6 +3,7 @@ exclude = [
     "debug_utils/casm-data-flow",
 ]
 members = [
+    "cairo-starknet-syscalls",
     "debug_utils",
     "debug_utils/sierra-emu",
     "debug_utils/cairo-native-stress",
@@ -53,6 +54,7 @@ cairo-lang-test-runner = "~2.19.0-rc.1"
 cairo-lang-utils = "~2.19.0-rc.1"
 cairo-native-bin-utils.path = "binaries/cairo-native-bin-utils"
 cairo-native = { path = ".", version = "0.9.0-rc.7" }
+cairo-starknet-syscalls = { path = "cairo-starknet-syscalls", version = "0.9.0-rc.7" }
 clap = "4.5.23"
 colored = "2.1.0"
 criterion = "0.5.1"
@@ -135,6 +137,7 @@ testing = ["dep:cairo-lang-compiler", "dep:cairo-lang-filesystem"]
 [dependencies]
 aquamarine.workspace = true
 bumpalo.workspace = true
+cairo-starknet-syscalls.workspace = true
 cairo-lang-compiler = { workspace = true, optional = true }
 cairo-lang-filesystem = { workspace = true, optional = true }
 cairo-lang-lowering.workspace = true
