+++ cairo-starknet-syscalls/src/lib.rs
@@ -1,0 +1,391 @@
+//! Shared `StarknetSyscallHandler` trait and supporting types for cairo-native
+//! and sierra-emu.
+//!
+//! Both crates re-export from here so a single syscall-handler impl can drive
+//! both the cairo-native AOT executor and the sierra-emu interpreter.
+
+#![deny(unused_must_use)]
+
+use serde::{Deserialize, Serialize};
+use starknet_types_core::felt::Felt;
+
+pub type SyscallResult<T> = std::result::Result<T, Vec<Felt>>;
+
+/// Binary representation of a `u256` (in MLIR).
+#[derive(
+    Debug,
+    Clone,
+    Copy,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    serde::Serialize,
+    serde::Deserialize,
+    Default,
+)]
+#[repr(C, align(16))]
+pub struct U256 {
+    pub lo: u128,
+    pub hi: u128,
+}
+
+#[derive(
+    Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    Default,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct ExecutionInfo {
+    pub block_info: BlockInfo,
+    pub tx_info: TxInfo,
+    pub caller_address: Felt,
+    pub contract_address: Felt,
+    pub entry_point_selector: Felt,
+}
+
+#[derive(
+    Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, serde::Serialize, serde::Deserialize,
+)]
+pub struct ExecutionInfoV2 {
+    pub block_info: BlockInfo,
+    pub tx_info: TxV2Info,
+    pub caller_address: Felt,
+    pub contract_address: Felt,
+    pub entry_point_selector: Felt,
+}
+
+#[derive(
+    Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    Default,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct TxV2Info {
+    pub version: Felt,
+    pub account_contract_address: Felt,
+    pub max_fee: u128,
+    pub signature: Vec<Felt>,
+    pub transaction_hash: Felt,
+    pub chain_id: Felt,
+    pub nonce: Felt,
+    pub resource_bounds: Vec<ResourceBounds>,
+    pub tip: u128,
+    pub paymaster_data: Vec<Felt>,
+    pub nonce_data_availability_mode: u32,
+    pub fee_data_availability_mode: u32,
+    pub account_deployment_data: Vec<Felt>,
+}
+
+#[derive(
+    Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, serde::Serialize, serde::Deserialize,
+)]
+pub struct ExecutionInfoV3 {
+    pub block_info: BlockInfo,
+    pub tx_info: TxV3Info,
+    pub caller_address: Felt,
+    pub contract_address: Felt,
+    pub entry_point_selector: Felt,
+}
+
+#[derive(
+    Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    Default,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct TxV3Info {
+    pub version: Felt,
+    pub account_contract_address: Felt,
+    pub max_fee: u128,
+    pub signature: Vec<Felt>,
+    pub transaction_hash: Felt,
+    pub chain_id: Felt,
+    pub nonce: Felt,
+    pub resource_bounds: Vec<ResourceBounds>,
+    pub tip: u128,
+    pub paymaster_data: Vec<Felt>,
+    pub nonce_data_availability_mode: u32,
+    pub fee_data_availability_mode: u32,
+    pub account_deployment_data: Vec<Felt>,
+    pub proof_facts: Vec<Felt>,
+}
+
+#[derive(
+    Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, serde::Serialize, serde::Deserialize,
+)]
+pub struct ResourceBounds {
+    pub resource: Felt,
+    pub max_amount: u64,
+    pub max_price_per_unit: u128,
+}
+
+#[derive(
+    Debug,
+    Clone,
+    Copy,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    Default,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct BlockInfo {
+    pub block_number: u64,
+    pub block_timestamp: u64,
+    pub sequencer_address: Felt,
+}
+
+#[derive(
+    Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    PartialOrd,
+    Ord,
+    Hash,
+    Default,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct TxInfo {
+    pub version: Felt,
+    pub account_contract_address: Felt,
+    pub max_fee: u128,
+    pub signature: Vec<Felt>,
+    pub transaction_hash: Felt,
+    pub chain_id: Felt,
+    pub nonce: Felt,
+}
+
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Deserialize, Serialize, Default)]
+#[repr(C, align(16))]
+pub struct Secp256k1Point {
+    pub x: U256,
+    pub y: U256,
+    pub is_infinity: bool,
+}
+
+impl Secp256k1Point {
+    pub const fn new(x_lo: u128, x_hi: u128, y_lo: u128, y_hi: u128, is_infinity: bool) -> Self {
+        Self {
+            x: U256 { lo: x_lo, hi: x_hi },
+            y: U256 { lo: y_lo, hi: y_hi },
+            is_infinity,
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Deserialize, Serialize, Default)]
+#[repr(C, align(16))]
+pub struct Secp256r1Point {
+    pub x: U256,
+    pub y: U256,
+    pub is_infinity: bool,
+}
+
+impl Secp256r1Point {
+    pub const fn new(x_lo: u128, x_hi: u128, y_lo: u128, y_hi: u128, is_infinity: bool) -> Self {
+        Self {
+            x: U256 { lo: x_lo, hi: x_hi },
+            y: U256 { lo: y_lo, hi: y_hi },
+            is_infinity,
+        }
+    }
+}
+
+pub trait StarknetSyscallHandler {
+    fn get_block_hash(&mut self, block_number: u64, remaining_gas: &mut u64)
+        -> SyscallResult<Felt>;
+
+    fn get_execution_info(&mut self, remaining_gas: &mut u64) -> SyscallResult<ExecutionInfo>;
+
+    fn get_execution_info_v2(&mut self, remaining_gas: &mut u64) -> SyscallResult<ExecutionInfoV2>;
+
+    fn get_execution_info_v3(&mut self, remaining_gas: &mut u64) -> SyscallResult<ExecutionInfoV3>;
+
+    fn deploy(
+        &mut self,
+        class_hash: Felt,
+        contract_address_salt: Felt,
+        calldata: &[Felt],
+        deploy_from_zero: bool,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<(Felt, Vec<Felt>)>;
+
+    fn replace_class(&mut self, class_hash: Felt, remaining_gas: &mut u64) -> SyscallResult<()>;
+
+    fn library_call(
+        &mut self,
+        class_hash: Felt,
+        function_selector: Felt,
+        calldata: &[Felt],
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Vec<Felt>>;
+
+    fn call_contract(
+        &mut self,
+        address: Felt,
+        entry_point_selector: Felt,
+        calldata: &[Felt],
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Vec<Felt>>;
+
+    fn storage_read(
+        &mut self,
+        address_domain: u32,
+        address: Felt,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Felt>;
+
+    fn storage_write(
+        &mut self,
+        address_domain: u32,
+        address: Felt,
+        value: Felt,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<()>;
+
+    fn emit_event(
+        &mut self,
+        keys: &[Felt],
+        data: &[Felt],
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<()>;
+
+    fn send_message_to_l1(
+        &mut self,
+        to_address: Felt,
+        payload: &[Felt],
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<()>;
+
+    fn keccak(&mut self, input: &[u64], remaining_gas: &mut u64) -> SyscallResult<U256>;
+
+    fn secp256k1_new(
+        &mut self,
+        x: U256,
+        y: U256,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Option<Secp256k1Point>>;
+
+    fn secp256k1_add(
+        &mut self,
+        p0: Secp256k1Point,
+        p1: Secp256k1Point,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Secp256k1Point>;
+
+    fn secp256k1_mul(
+        &mut self,
+        p: Secp256k1Point,
+        m: U256,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Secp256k1Point>;
+
+    fn secp256k1_get_point_from_x(
+        &mut self,
+        x: U256,
+        y_parity: bool,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Option<Secp256k1Point>>;
+
+    fn secp256k1_get_xy(
+        &mut self,
+        p: Secp256k1Point,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<(U256, U256)>;
+
+    fn secp256r1_new(
+        &mut self,
+        x: U256,
+        y: U256,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Option<Secp256r1Point>>;
+
+    fn secp256r1_add(
+        &mut self,
+        p0: Secp256r1Point,
+        p1: Secp256r1Point,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Secp256r1Point>;
+
+    fn secp256r1_mul(
+        &mut self,
+        p: Secp256r1Point,
+        m: U256,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Secp256r1Point>;
+
+    fn secp256r1_get_point_from_x(
+        &mut self,
+        x: U256,
+        y_parity: bool,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Option<Secp256r1Point>>;
+
+    fn secp256r1_get_xy(
+        &mut self,
+        p: Secp256r1Point,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<(U256, U256)>;
+
+    fn sha256_process_block(
+        &mut self,
+        state: &mut [u32; 8],
+        block: &[u32; 16],
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<()>;
+
+    fn sha512_process_block(
+        &mut self,
+        state: &mut [u64; 8],
+        block: &[u64; 16],
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<()>;
+
+    fn get_class_hash_at(
+        &mut self,
+        contract_address: Felt,
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Felt>;
+
+    fn meta_tx_v0(
+        &mut self,
+        address: Felt,
+        entry_point_selector: Felt,
+        calldata: &[Felt],
+        signature: &[Felt],
+        remaining_gas: &mut u64,
+    ) -> SyscallResult<Vec<Felt>>;
+
+    /// Test-only Starknet syscall. Production handlers don't implement it; the default
+    /// returns a single error felt rather than panicking, so a malicious contract that
+    /// invokes `cheatcode` against a handler that didn't override the method can't crash
+    /// the host. Test handlers should override.
+    fn cheatcode(&mut self, _selector: Felt, _input: &[Felt]) -> Vec<Felt> {
+        vec![Felt::from_bytes_be_slice(b"cheatcode unsupported")]
+    }
+}
