--- debug_utils/sierra-emu/src/starknet/execution_info_v2.rs
@@ -1,30 +1,0 @@
-use super::{BlockInfo, TxV2Info};
-use crate::Value;
-use cairo_lang_sierra::ids::ConcreteTypeId;
-use serde::Serialize;
-use starknet_types_core::felt::Felt;
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize)]
-pub struct ExecutionInfoV2 {
-    pub block_info: BlockInfo,
-    pub tx_info: TxV2Info,
-    pub caller_address: Felt,
-    pub contract_address: Felt,
-    pub entry_point_selector: Felt,
-}
-
-impl ExecutionInfoV2 {
-    pub(crate) fn into_value(
-        self,
-        felt252_ty: ConcreteTypeId,
-        resource_bounds_ty: ConcreteTypeId,
-    ) -> Value {
-        Value::Struct(vec![
-            self.block_info.into_value(),
-            self.tx_info.into_value(felt252_ty, resource_bounds_ty),
-            Value::Felt(self.caller_address),
-            Value::Felt(self.contract_address),
-            Value::Felt(self.entry_point_selector),
-        ])
-    }
-}
