--- debug_utils/sierra-emu/src/starknet/tx_v2_info.rs
@@ -1,66 +1,0 @@
-use super::ResourceBounds;
-use crate::Value;
-use cairo_lang_sierra::ids::ConcreteTypeId;
-use serde::Serialize;
-use starknet_types_core::felt::Felt;
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize)]
-pub struct TxV2Info {
-    pub version: Felt,
-    pub account_contract_address: Felt,
-    pub max_fee: u128,
-    pub signature: Vec<Felt>,
-    pub transaction_hash: Felt,
-    pub chain_id: Felt,
-    pub nonce: Felt,
-    pub resource_bounds: Vec<ResourceBounds>,
-    pub tip: u128,
-    pub paymaster_data: Vec<Felt>,
-    pub nonce_data_availability_mode: u32,
-    pub fee_data_availability_mode: u32,
-    pub account_deployment_data: Vec<Felt>,
-}
-
-impl TxV2Info {
-    pub(crate) fn into_value(
-        self,
-        felt252_ty: ConcreteTypeId,
-        resource_bounds_ty: ConcreteTypeId,
-    ) -> Value {
-        Value::Struct(vec![
-            Value::Felt(self.version),
-            Value::Felt(self.account_contract_address),
-            Value::U128(self.max_fee),
-            Value::Struct(vec![Value::Array {
-                ty: felt252_ty.clone(),
-                data: self.signature.into_iter().map(Value::Felt).collect(),
-            }]),
-            Value::Felt(self.transaction_hash),
-            Value::Felt(self.chain_id),
-            Value::Felt(self.nonce),
-            Value::Struct(vec![Value::Array {
-                ty: resource_bounds_ty,
-                data: self
-                    .resource_bounds
-                    .into_iter()
-                    .map(ResourceBounds::into_value)
-                    .collect(),
-            }]),
-            Value::U128(self.tip),
-            Value::Struct(vec![Value::Array {
-                ty: felt252_ty.clone(),
-                data: self.paymaster_data.into_iter().map(Value::Felt).collect(),
-            }]),
-            Value::U32(self.nonce_data_availability_mode),
-            Value::U32(self.fee_data_availability_mode),
-            Value::Struct(vec![Value::Array {
-                ty: felt252_ty,
-                data: self
-                    .account_deployment_data
-                    .into_iter()
-                    .map(Value::Felt)
-                    .collect(),
-            }]),
-        ])
-    }
-}
