+++ debug_utils/sierra-emu/src/vm/starknet.rs
@@ -1,6 +1,13 @@
 use super::EvalAction;
 use crate::{
-    starknet::{Secp256k1Point, Secp256r1Point, StarknetSyscallHandler, U256},
+    starknet::{
+        value_conv::{
+            execution_info_into_value, execution_info_v2_into_value, secp256k1_point_from_value,
+            secp256k1_point_into_value, secp256r1_point_from_value, secp256r1_point_into_value,
+            u256_from_value, u256_into_value,
+        },
+        StarknetSyscallHandler, U256,
+    },
     Value,
 };
 use cairo_lang_sierra::{
@@ -203,10 +210,10 @@ fn eval_secp256_new(
     let syscall_result = match secp_type {
         SecpPointType::K1 => syscall_handler
             .secp256k1_new(x, y, &mut gas)
-            .map(|res| res.map(|op| op.into_value())),
+            .map(|res| res.map(secp256k1_point_into_value)),
         SecpPointType::R1 => syscall_handler
             .secp256r1_new(x, y, &mut gas)
-            .map(|res| res.map(|op| op.into_value())),
+            .map(|res| res.map(secp256r1_point_into_value)),
     };
 
     match syscall_result {
@@ -269,20 +276,20 @@ fn eval_secp256_add(
 
     let syscall_result = match secp_type {
         SecpPointType::K1 => {
-            let x = Secp256k1Point::from_value(x);
-            let y = Secp256k1Point::from_value(y);
+            let x = secp256k1_point_from_value(x);
+            let y = secp256k1_point_from_value(y);
 
             syscall_handler
                 .secp256k1_add(x, y, &mut gas)
-                .map(|res| res.into_value())
+                .map(secp256k1_point_into_value)
         }
         SecpPointType::R1 => {
-            let x = Secp256r1Point::from_value(x);
-            let y = Secp256r1Point::from_value(y);
+            let x = secp256r1_point_from_value(x);
+            let y = secp256r1_point_from_value(y);
 
             syscall_handler
                 .secp256r1_add(x, y, &mut gas)
-                .map(|res| res.into_value())
+                .map(secp256r1_point_into_value)
         }
     };
 
@@ -307,22 +314,22 @@ fn eval_secp256_mul(
         panic!()
     };
 
-    let n = U256::from_value(n);
+    let n = u256_from_value(n);
 
     let syscall_result = match secp_type {
         SecpPointType::K1 => {
-            let x = Secp256k1Point::from_value(x);
+            let x = secp256k1_point_from_value(x);
 
             syscall_handler
                 .secp256k1_mul(x, n, &mut gas)
-                .map(|res| res.into_value())
+                .map(secp256k1_point_into_value)
         }
         SecpPointType::R1 => {
-            let x = Secp256r1Point::from_value(x);
+            let x = secp256r1_point_from_value(x);
 
             syscall_handler
                 .secp256r1_mul(x, n, &mut gas)
-                .map(|res| res.into_value())
+                .map(secp256r1_point_into_value)
         }
     };
 
@@ -357,10 +364,10 @@ fn eval_secp256_get_point_from_x(
     let syscall_result = match secp_type {
         SecpPointType::K1 => syscall_handler
             .secp256k1_get_point_from_x(x, y_parity, &mut gas)
-            .map(|res| res.map(|op| op.into_value())),
+            .map(|res| res.map(secp256k1_point_into_value)),
         SecpPointType::R1 => syscall_handler
             .secp256r1_get_point_from_x(x, y_parity, &mut gas)
-            .map(|res| res.map(|op| op.into_value())),
+            .map(|res| res.map(secp256r1_point_into_value)),
     };
 
     match syscall_result {
@@ -425,14 +432,14 @@ fn eval_secp256_get_xy(
 
     let syscall_result = match secp_type {
         SecpPointType::K1 => {
-            let secp_value = Secp256k1Point::from_value(secp_value);
+            let secp_value = secp256k1_point_from_value(secp_value);
 
             syscall_handler
                 .secp256k1_get_xy(secp_value, &mut gas)
                 .map(|res| (res.0, res.1))
         }
         SecpPointType::R1 => {
-            let secp_value = Secp256r1Point::from_value(secp_value);
+            let secp_value = secp256r1_point_from_value(secp_value);
 
             syscall_handler
                 .secp256r1_get_xy(secp_value, &mut gas)
@@ -442,7 +449,7 @@ fn eval_secp256_get_xy(
 
     match syscall_result {
         Ok(payload) => {
-            let (x, y) = (payload.0.into_value(), payload.1.into_value());
+            let (x, y) = (u256_into_value(payload.0), u256_into_value(payload.1));
             EvalAction::NormalBranch(0, smallvec![Value::U64(gas), system, x, y])
         }
         Err(payload) => {
@@ -903,7 +910,11 @@ fn eval_get_execution_info(
     match result {
         Ok(res) => EvalAction::NormalBranch(
             0,
-            smallvec![Value::U64(gas), system, res.into_value(felt_ty)],
+            smallvec![
+                Value::U64(gas),
+                system,
+                execution_info_into_value(res, felt_ty)
+            ],
         ),
         Err(e) => EvalAction::NormalBranch(
             1,
@@ -984,7 +995,7 @@ fn eval_get_execution_info_v2(
             smallvec![
                 Value::U64(gas),
                 system,
-                res.into_value(felt_ty, out_ty_id.clone())
+                execution_info_v2_into_value(res, felt_ty, out_ty_id.clone())
             ],
         ),
         Err(e) => EvalAction::NormalBranch(
@@ -1121,7 +1132,7 @@ fn eval_keccak(
 
     match result {
         Ok(res) => {
-            EvalAction::NormalBranch(0, smallvec![Value::U64(gas), system, res.into_value()])
+            EvalAction::NormalBranch(0, smallvec![Value::U64(gas), system, u256_into_value(res)])
         }
         Err(e) => EvalAction::NormalBranch(
             1,
@@ -1420,7 +1431,7 @@ fn eval_sha512_process_block(
         panic!()
     };
 
-    let prev_state: [u64; 8] = prev_state
+    let mut prev_state: [u64; 8] = prev_state
         .into_iter()
         .map(|v| {
             let Value::U64(v) = v else { panic!() };
@@ -1450,9 +1461,9 @@ fn eval_sha512_process_block(
         }
     };
 
-    match syscall_handler.sha512_process_block(prev_state, current_block, &mut gas) {
-        Ok(payload) => {
-            let payload = payload.into_iter().map(Value::U64).collect::<Vec<_>>();
+    match syscall_handler.sha512_process_block(&mut prev_state, &current_block, &mut gas) {
+        Ok(()) => {
+            let payload = prev_state.into_iter().map(Value::U64).collect::<Vec<_>>();
             EvalAction::NormalBranch(
                 0,
                 smallvec![Value::U64(gas), system, Value::Struct(payload)],
